@@ -34,14 +34,14 @@
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import { defineComponent } from "vue";
 
 type DataType = {
   keyPhrase: string;
   phraseHistory: string[];
 };
 
-export default Vue.extend({
+export default defineComponent({
   name: "EditorOption",
   data(): DataType {
     const phraseHistory: string[] = JSON.parse(
