@@ -5,10 +5,10 @@
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import { defineComponent } from "vue";
 import ConfigureMain from "@/components/configure/ConfigureMain.vue";
 
-export default Vue.extend({
+export default defineComponent({
   name: "Configure",
   components: {
     ConfigureMain
